feat: update dependencies and Docker configurations 🚀
This commit updates several dependencies and Dockerfile configurations to their latest versions for improved performance and security. Keeping dependencies up-to-date is crucial for maintaining the stability and security of the application.

- Updated Git version to `1:2.51.0` in `Dockerfile.desktop` and `Dockerfile.shell`
- Upgraded Go to version `1.24.5`
- Changed the dotfiles repo version to `v1.120.8`
- Updated UV to `0.8.3`
- Upgraded Node.js version to `24.4.1`
- Migrated `diff-so-fancy` to `delta` and upgraded accordingly
- Updated `Rust` to `1.88.0`
- Upgraded `sudo` version to `1.9.15p5-3ubuntu5.24.04.1`
- Updated multiple tools including `jq`, `gh`, and `lazygit` to newer versions
- Introduced several new components such as `sqlite-utils`, `docker-compose`, and `container-diff`

The goal of these changes is to improve compatibility and take advantage of the latest features and security patches offered by the dependencies.